chore: bump OT core to v0.45.0
