+++ pkg/exporter/sumologicexporter/carbon_formatter.go
@@ -66,7 +66,7 @@ func sanitizeCarbonString(text string) string {
 // carbon2NumberRecord converts NumberDataPoint to carbon2 metric string
 // with additional information from metricPair.
 func carbon2NumberRecord(record metricPair, dataPoint pdata.NumberDataPoint) string {
-	switch dataPoint.Type() {
+	switch dataPoint.ValueType() {
 	case pdata.MetricValueTypeDouble:
 		return fmt.Sprintf("%s  %g %d",
 			carbon2TagString(record),
