+++ pkg/exporter/sumologicexporter/exporter.go
@@ -60,47 +60,6 @@ type sumologicexporter struct {
 }
 
 func initExporter(cfg *Config, createSettings component.ExporterCreateSettings) (*sumologicexporter, error) {
-	switch cfg.LogFormat {
-	case JSONFormat:
-	case TextFormat:
-	case OTLPLogFormat:
-	default:
-		return nil, fmt.Errorf("unexpected log format: %s", cfg.LogFormat)
-	}
-
-	switch cfg.MetricFormat {
-	case GraphiteFormat:
-	case Carbon2Format:
-	case PrometheusFormat:
-	case OTLPMetricFormat:
-	default:
-		return nil, fmt.Errorf("unexpected metric format: %s", cfg.MetricFormat)
-	}
-
-	switch cfg.TraceFormat {
-	case OTLPTraceFormat:
-	default:
-		return nil, fmt.Errorf("unexpected trace format: %s", cfg.TraceFormat)
-	}
-
-	switch cfg.CompressEncoding {
-	case GZIPCompression:
-	case DeflateCompression:
-	case NoCompression:
-	default:
-		return nil, fmt.Errorf("unexpected compression encoding: %s", cfg.CompressEncoding)
-	}
-
-	if len(cfg.HTTPClientSettings.Endpoint) == 0 && cfg.HTTPClientSettings.Auth == nil {
-		return nil, errors.New("no endpoint and no auth extension specified")
-	}
-
-	if _, err := url.Parse(cfg.HTTPClientSettings.Endpoint); err != nil {
-		return nil, fmt.Errorf("failed parsing endpoint URL: %s; err: %w",
-			cfg.HTTPClientSettings.Endpoint, err,
-		)
-	}
-
 	if cfg.TranslateAttributes {
 		cfg.SourceCategory = translateConfigValue(cfg.SourceCategory)
 		cfg.SourceHost = translateConfigValue(cfg.SourceHost)
