+++ pkg/exporter/sumologicexporter/exporter_test.go
@@ -131,79 +131,6 @@ func TestInitExporter(t *testing.T) {
 	assert.NoError(t, err)
 }
 
-func TestInitExporterInvalidLogFormat(t *testing.T) {
-	_, err := initExporter(&Config{
-		LogFormat:        "test_format",
-		MetricFormat:     "carbon2",
-		CompressEncoding: "gzip",
-		TraceFormat:      "otlp",
-		HTTPClientSettings: confighttp.HTTPClientSettings{
-			Timeout:  defaultTimeout,
-			Endpoint: "test_endpoint",
-		},
-	}, createExporterCreateSettings())
-
-	assert.EqualError(t, err, "unexpected log format: test_format")
-}
-
-func TestInitExporterInvalidMetricFormat(t *testing.T) {
-	_, err := initExporter(&Config{
-		LogFormat:    "json",
-		MetricFormat: "test_format",
-		HTTPClientSettings: confighttp.HTTPClientSettings{
-			Timeout:  defaultTimeout,
-			Endpoint: "test_endpoint",
-		},
-		CompressEncoding: "gzip",
-	}, createExporterCreateSettings())
-
-	assert.EqualError(t, err, "unexpected metric format: test_format")
-}
-
-func TestInitExporterInvalidTraceFormat(t *testing.T) {
-	_, err := initExporter(&Config{
-		LogFormat:    "json",
-		MetricFormat: "carbon2",
-		TraceFormat:  "text",
-		HTTPClientSettings: confighttp.HTTPClientSettings{
-			Timeout:  defaultTimeout,
-			Endpoint: "test_endpoint",
-		},
-		CompressEncoding: "gzip",
-	}, createExporterCreateSettings())
-
-	assert.EqualError(t, err, "unexpected trace format: text")
-}
-
-func TestInitExporterInvalidCompressEncoding(t *testing.T) {
-	_, err := initExporter(&Config{
-		LogFormat:        "json",
-		MetricFormat:     "carbon2",
-		CompressEncoding: "test_format",
-		TraceFormat:      "otlp",
-		HTTPClientSettings: confighttp.HTTPClientSettings{
-			Timeout:  defaultTimeout,
-			Endpoint: "test_endpoint",
-		},
-	}, createExporterCreateSettings())
-
-	assert.EqualError(t, err, "unexpected compression encoding: test_format")
-}
-
-func TestInitExporterInvalidEndpointAndNoAuth(t *testing.T) {
-	_, err := initExporter(&Config{
-		LogFormat:        "json",
-		MetricFormat:     "carbon2",
-		CompressEncoding: "gzip",
-		TraceFormat:      "otlp",
-		HTTPClientSettings: confighttp.HTTPClientSettings{
-			Timeout: defaultTimeout,
-		},
-	}, createExporterCreateSettings())
-
-	assert.EqualError(t, err, "no endpoint and no auth extension specified")
-}
-
 func TestAllSuccess(t *testing.T) {
 	test := prepareExporterTest(t, createTestConfig(), []func(w http.ResponseWriter, req *http.Request){
 		func(w http.ResponseWriter, req *http.Request) {
